+++ src/Database/V5/Bloodhound/Internal/Analysis.hs
@@ -5,11 +5,7 @@ module Database.V5.Bloodhound.Internal.Analysis where
 
 import           Bloodhound.Import
 
-import           Data.Aeson
-import           Data.Aeson.Types (Parser)
 import qualified Data.Map.Strict as M
-import           Data.Maybe         (catMaybes)
-import           Data.Text          (Text)
 import qualified Data.Text as T
 
 import           Database.V5.Bloodhound.Internal.Newtypes
@@ -216,7 +212,7 @@ data Language
   deriving (Eq, Show)
 
 instance ToJSON Language where
-  toJSON = Data.Aeson.String . languageToText
+  toJSON = String . languageToText
 
 instance FromJSON Language where
   parseJSON = withText "Language" $ \t -> case languageFromText t of
